requirements/base.txt
Babel==2.3.4
blessings==1.6
bpython==0.15
curtsies==0.2.6
defusedxml==0.4.1
dj-database-url==0.4.0
dj-static==0.0.6
Django==1.9.2
django-allauth==0.24.1
django-braces==1.8.1
django-countries==3.4.1
django-crispy-forms==1.6.0
django-debug-toolbar==1.4
django-environ==0.4.0
django-extensions==1.6.1
django-floppyforms==1.6.1
django-mailgun==0.8.0
django-modeltranslation==0.11
django-parler==1.6.5
django-paypal==0.3.2
django-phonenumber-field==1.1.0
django-rosetta==0.7.11
easy-thumbnails==2.3
greenlet==0.4.9
gunicorn==19.4.5
markdown2==2.3.1
microsofttranslator==0.8
mistune==0.7.2
oauthlib==1.1.2
phonenumberslite==7.5.1
Pillow==3.1.1
pluggy==0.3.1
polib==1.0.7
psycopg2==2.6.1
py==1.4.31
Pygments==2.1.3
python3-openid==3.0.10
pytz==2016.6.1
qrcode==5.3
redis==2.10.5
reportlab==3.3.0
requests==2.9.1
requests-oauthlib==0.6.2
SimpleCV==1.3
six==1.10.0
sqlparse==0.1.19
static3==0.7.0
svg2rlg==0.3
tox==2.3.1
virtualenv==15.0.3
wcwidth==0.1.6
Werkzeug==0.11.5
whitenoise==2.0.6
